refactor : utils 폴더 업데이트
